ml-service/ml-service/requirements.txt
fastapi
uvicorn
pandas
requests
anthropic
python-dotenv